move activations tests from staff_controller_test to activations_controller_test
